backend/requirements.txt
flask==2.3.2
tensorflow-cpu==2.12.0
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
joblib==1.3.2
gunicorn==20.1.0
numpy==1.23.5